wip: enhance service
