+++ app/src/services/tipsService.ts
@@ -3,42 +3,15 @@ import { TIPS_ADDRESS } from '@/constant'
 import ABI from '../artifacts/abi/tips.json'
 import type { TipsEventType } from '@/types'
 import type { EventLog } from 'ethers'
-import type { Contract } from 'ethers'
 import type { Log } from 'ethers'
 import { BankService, type IBankService } from './bankService'
+import { SmartContract } from './contractService'
 
-interface ISmartContract {
-  contract: Contract | undefined
-  contractAddress: string
-  contractAbi: any
-  getContract(address: string, abi: any): Promise<Contract>
-}
-
-export class SmartContract implements ISmartContract {
-  contract: Contract | undefined
-  contractAddress: string
-  contractAbi: any
-  web3Library: IWeb3Library
-
-  constructor(web3Library: IWeb3Library, contractAddress: string, contractAbi: any) {
-    this.web3Library = web3Library
-    this.contractAddress = contractAddress
-    this.contractAbi = contractAbi
-  }
-
-  async getContract(): Promise<Contract> {
-    if (!this.contract) {
-      this.contract = await this.web3Library.getContract(this.contractAddress, this.contractAbi)
-    }
-
-    return this.contract
-  }
-}
 
 export class TipsService extends SmartContract {
   bankService: IBankService
   constructor(web3Library: IWeb3Library = EthersJsAdapter.getInstance()) {
-    super(web3Library, TIPS_ADDRESS, ABI)
+    super(TIPS_ADDRESS, ABI, web3Library)
     this.bankService = new BankService()
   }
 
@@ -94,13 +67,8 @@ export class TipsService extends SmartContract {
     await tx.wait()
   }
 
-  // TODO use a builder pattern for this
   async getEvents(type: TipsEventType): Promise<EventLog[] | Log[]> {
-    if (!this.contract) {
-      this.contract = await super.getContract()
-    }
-
-    return this.contract.queryFilter(type)
+    return await super.getEvents(type)
   }
 
   async bankPushTip(bankAddress: string, addresses: string[], amount: number): Promise<any> {
